requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
isodate==0.6.0
mysql-connector==2.2.9
requests==2.22.0
six==1.14.0
urllib3==1.25.8
